+++ src/pages/community/CreateCommunity/findRequest/styles.tsx
