+++ src/pages/community/CreateCommunity/question/components/setVoteTime/styles.tsx
