+++ src/pages/community/pages/FindHome/FindHome.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const FindHome = () => {
+  return <div>FindHome</div>
+}
+
+export default FindHome
