+++ src/pages/community/pages/HowAboutHome/HowAboutHome.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const HowAboutHome = () => {
+  return <div>HowAboutHome</div>
+}
+
+export default HowAboutHome
