+++ src/pages/community/pages/RecommendHome/RecommendHome.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const RecommendHome = () => {
+  return <div>RecommendHome</div>
+}
+
+export default RecommendHome
