+++ src/pages/community/pages/WhichItemHome/WhichItemHome.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const WhichItemHome = () => {
+  return <div>WhichItemHome</div>
+}
+
+export default WhichItemHome
